+++ sketch_oxide/src/statistics/mod.rs
@@ -6,8 +6,10 @@
 
 mod ams;
 mod hll_joint;
+mod k_ary;
 mod p_stable;
 
 pub use ams::AmsSketch;
 pub use hll_joint::HllJointEstimator;
+pub use k_ary::KArySketch;
 pub use p_stable::PStableLpSketch;